Custom request examples misleading
I think this part is wrong or at least misleading. If I'm mistaking, please add a bit more of documentation to explain how that works. Beginners like may have hard times with this. ;)

Well I wanted to use the custom put and took for example the customPOST from the let's code section.

It says here that it is going to do a POST /accounts/123/messages?param=myParam but it is actually doing more of a POST /accounts/123/[Object object]?param=MyParam

I solved this by adding an empty object as first param. That seems coherent with the documentation below. the element appears before the path:
customPOST([elem, path, params, headers])

Now it also says that elem is optionnal but apparently not if there are params provided (may be that's the part to update). Anyway, I think these examples of custom http verbs are not accurate. :)

Thanks for restangular anyway, It's great! :D